requirements.txt
flask
flask_cors
apscheduler
scrapy
crochet
pymongo
python-dotenv

